TODO: this could be done as a runtime bool() rather than at the type level -- there's nothing actually needed for type verification in the marker class at all - all the marker is doing is avoiding thready/race condidiotns and semantic meaninglessness of shared processes.
Disable monitoring wrapper for non-htex

there are different choices i could implement:

 per-executor impl whitelisting of monitor wrapper (controlled by devs)
 per-executor config of monitor wrapper (more burden on users, more configurable path for future
options, worse config orthogonality - which is a problem in general with 2d config)

Whats the problem?

Conceptually, the monitoring wrapper doesn't work for executors which run simultaneous app invocations in the same process.

More practically, the montoring wrapper uses process launching which is explicitly stated to not work alongside threaded
stuff - (there's a ref for this as a mention in some python doc, which should go in this commit message) and so this hangs
things sometimes with a deadlock.

So this should only be enabled for executors which run tasks in separate processes.

But I probably still want the (non-threaded) start message that gives the running status in the monitoring db (distinct from resource usage)
even if this is disabled.

This PR should mean that monitoring can be enabled with threadpoolexecutor - a pre-req for join apps, and incidentally a requirement for globus staging vs monitoring

This should (but doesn't?) make task_try_time_running appear for local executor tasks too - because the *first* message shoudl always be sent (not from a separate process)

+++ parsl/monitoring/monitoring.py
@@ -388,7 +388,8 @@ class MonitoringHub(RepresentationMixin):
                         run_id: str,
                         logging_level: int,
                         sleep_dur: float,
-                        radio_mode: str) -> Callable:
+                        radio_mode: str,
+                        monitor_resources: bool) -> Callable:
         """ Internal
         Wrap the Parsl app with a function that will call the monitor function and point it at the correct pid when the task begins.
         """
@@ -400,25 +401,30 @@ class MonitoringHub(RepresentationMixin):
                                run_id,
                                radio_mode)
 
-            # create the monitor process and start
-            p = Process(target=monitor,
-                        args=(os.getpid(),
-                              try_id,
-                              task_id,
-                              monitoring_hub_url,
-                              run_id,
-                              radio_mode,
-                              logging_level,
-                              sleep_dur),
-                        name="Monitor-Wrapper-{}".format(task_id))
-            p.start()
+            if monitor_resources:
+                # create the monitor process and start
+                p: Optional[Process]
+                p = Process(target=monitor,
+                            args=(os.getpid(),
+                                  try_id,
+                                  task_id,
+                                  monitoring_hub_url,
+                                  run_id,
+                                  radio_mode,
+                                  logging_level,
+                                  sleep_dur),
+                            name="Monitor-Wrapper-{}".format(task_id))
+                p.start()
+            else:
+                p = None
 
             try:
                 return f(*args, **kwargs)
             finally:
                 # There's a chance of zombification if the workers are killed by some signals
-                p.terminate()
-                p.join()
+                if p:
+                    p.terminate()
+                    p.join()
         return wrapped
